# pkg/encryption: don't fragment read sizes so much