with this dumb local-only benchmark:
```
old: BenchmarkTransformReader-4  122  9464527 ns/op
new: BenchmarkTransformReader-4  145  8873115 ns/op
```

currently, Read calls max out at a buffer size of the unencrypted block size, which appears to be
29 (rs min) * 256 (share size) - 16 (secretbox overhead), so, ~7k. It'd be great if we tried to keep read buffers large (16k-32k) so we made less syscalls.

this change changes the transformer to make large read calls (via the bufio.NewReader wrapper), and return large reads if possible (via the new Read wrapper).

Please describe the tests: new benchmark
 
Please describe the performance impact: about a millisecond faster for copying 10MB from local disk to local disk (9ms to 8ms)

## Code Review Checklist (to be filled out by reviewer)
 - [ ] NEW: Are there any Satellite database migrations? Are they forwards _and_ backwards compatible? 
 - [ ] Does the PR describe what changes are being made?
 - [ ] Does the PR describe why the changes are being made?
 - [ ] Does the code follow [our style guide](https://github.com/storj/docs/blob/master/code/Style.md)?
 - [ ] Does the code follow [our testing guide](https://github.com/storj/docs/blob/master/code/Testing.md)?
 - [ ] Is the PR appropriately sized? (If it could be broken into smaller PRs it should be)
 - [ ] Does the new code have enough tests? (*every* PR should have tests or justification otherwise. Bug-fix PRs especially)
 - [ ] Does the new code have enough documentation that answers "how do I use it?" and "what does it do?"? (both source documentation and [higher level](https://github.com/storj/docs), diagrams?)
 - [ ] Does any documentation need updating?
 - [ ] Do the database access patterns make sense?